sleep for at least 1s
The fuse of the integration logic pcb of the gripper blew when
the robot continuously sent commands there (which were apparently
executed with 400ms delay).

To ensure this does not happen again, increase the execution time to
a whole second. There should be no use-case here where the robot needs
to open/close more often than that.